+++ libs/cardano-ledger-core/src/Cardano/Ledger/Language.hs
@@ -1,0 +1,3 @@
+module Cardano.Ledger.Language (module Cardano.Ledger.Plutus.Language) where
+
+import Cardano.Ledger.Plutus.Language